Update `accountsChanged` event when logged out
The `accountsChanged` event would previously return `[undefined]` upon
logout (i.e. when `selectedAddress` was undefined).

Instead an empty array is now returned if the user is not logged in.
This seems more semantically correct, and is more in-line with
EIP-1193, which states that the event should emit an array of strings.