Use `img` elements for testing resource existence rather than `fetch`
The previous implementation of `resourceExists` was using the `same-origin`
mode for fetch but that breaks when a site uses another domain for their icons—
`static.example.com` or `images.example.com`, for example.

There isn't a fetch mode that works here: [1]

- `same-origin`: doesn't work for the reasons described
- `cors`: there's no guarantee that the 2nd origin in question is correctly
configured for CORS—a quick check on two sites confirmed this
- `navigate`: isn't applicable here
- `no-cors` hides the `status` and `statusText` so we can't use those

With that, this loads the resources as images and tests whether that works.
This will essentially be a `no-cors` request but with the status exposed
through `onload`+`onerror`, which will suffice for our use case.

  [1]:https://fetch.spec.whatwg.org/#concept-request-mode