Use mode:cors when checking for site icons
This change updates the fetch mode we use when requesting site icons to use `cors`
instead of `same-origin` as `same-origin` breaks when a site uses a different
domain for their icons, e.g. `static.example.com` or `images.example.com`. While
it is also possible that the 2nd origin here isn't correctly configured for CORS
that's a fewer sites breaking.