Skip emitting `data` event for `wallet_accountsChanged` (#73)
The `data` event was being used by some dapps (via `web3@1.x` for
example) in a way that we didn't anticpiate. Specifically these dapps
were expecting the payload for this event to contain a `params`
property, and they were breaking if this was absent.

We hadn't intended this event for public consumption in the first
place, but we _especially_ didn't mean for `wallet_accountsChanged` to
be sent to dapps. So for now the easiest solution is to not emit `data`
for these events.

Currently the only other notification sent from the extension is
`eth_subscription`, so the `data` event has been moved inside the block
that checks for that type. The `params` check was removed for being
redundant, since `eth_subscription` messages should always include a
`params` property.