Restore `data` provider event (#135)
This event is deprecated, and is not part of EIP-1193. However we did not plan to remove this; it was removed accidentally as part of #109. It should now behave exactly as it did in v6.3.0.

The text of the deprecation warning for this event has also been updated to provide a recommended alternative.