Fix warning on second `currentProvider` access
We expose the `currentProvider` property on our shim as a deprecated
way to access to provider, to make the task of migrating dapps away
from our injected web3 instance easier. The intent was to warn if that
property is used, but log an error to the console and inform MetaMask
if any other property is used.

There was a mistake in the logic that caused us to treat subsequent
accesses to `currentProvider` to be treated like some unsupported
property was being accessed. This results in the console error being
shown erroneously, it bloats our web3 usage metrics, and it causes the
warning about web3 usage to appear unnecessarily in the MetaMask UI.

The logic has been updated to never treat `currentProvider` as an
unsupported property. Accessing that property will only result in a
deprecation warning.