Move stream functionality from BaseProvider to new StreamProvider (#209)
The constructor of the `BaseProvider` expected a duplex stream, and the class as a whole contained a considerable amount of logic specific to streams and stream handling. This PR extracts this and other dependent functionality into a new class, `StreamProvider`, which is now extended by `MetaMaskInpageProvider`, while `BaseProvider` is made abstract. This will be useful for replacing internal legacy providers (which do not always / ever use streams) with EIP-1193 equivalents throughout the MetaMask codebase.

This PR should be completely non-breaking except for the changes made to the `BaseProvider` class. In addition, `createExternalExtensionProvider`, `MetaMaskInpageProvider`, and `initializeProvider` should have no behavior changes at all.

In detail, this PR:
- Converts `BaseProvider` to an abstract class, and moves all stream-specific functionality into a new class, `StreamProvider`
- Moves `BaseProvider` tests to the `StreamProvider` tests, adds an initialization test case
- Adds optional `rpcMiddleware` parameter to `BaseProvider`
  - Middlewares previously instantiated and pushed inside of `BaseProvider` are now instead passed as constructor parameters to `StreamProvider` and `MetaMaskInpageProvider` as necessary.
- Moves `networkVersion`-related logic from the `BaseProvider` to `MetaMaskInpageProvider`, while continuing to rely on the `networkVersion` having the value of `'loading'` to detect when we are mid-network change in the `StreamProvider` (and by extension its child classes).
- Updates the Jest config, adds `jest-it-up` and minimum coverage requirements