Remove protection against responding before init
A protection measure was setup in the initialized provider helper
function to prevent it from responding to messages before the
initialization was complete This has been removed.

It was originally added during debugging, to rule out before-
initialization messages as an explanation for a bug that was later
fixed I left it in because I thought that it might better match
expectations if messages before initialization were ignored.

But in retrospect, I think this would be both unnecessary and harmful.
It's unnecessary because for most tests, no messages would be sent
until after initialization anyway. And it's potentially harmful because
this doesn't represent how the provider actually behaves. This could
disguise a mistake in a test, or it could get in the way of testing
strange edge cases surrounding initialization.