feat: Modify provider RPC error log severity from `error` to `debug`
The error middleware injected in the provider-side `json-rpc-engine` was calling `log.error()` (in practice, `log` = `console`) every time it observed a JSON-RPC error response.

We received reports that this error log is caught by the monitoring infra of some dapps. It is ultimately the caller's responsibility to handle errors where they call `ethereum.request()`, and they can always `console.error()` at that point themselves. In light of the above, we decrease the severity of our own log statement to `debug`.