fix: Allow optional `jsonRpcStreamName` for inpage provider (#390)
* fix: Allow optional `jsonRpcStreamName` for inpage provider

The `MetaMaskInpageProvider` type was mistakenly typed to require the
`jsonRpcStreamName` as a constructor parameter, despite the author
clearly intending this to be optional. This type error was fixed.

This type error motivated a different bug, which is that validation for
`jsonRpcStreamName` was added to `initializeInapgeProvider` in #381,
constituting an undocumented breaking change (currently blocking us
from updating this package in `metamask-extension`). This validation
has been removed now that it's no longer needed.

Fixes #389

* Update test coverage